% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnt-package.R
\docType{package}
\name{acom-package}
\alias{acom}
\alias{acom-package}
\title{acom: Aphasia Communication Outcome Measure}
\description{
A Shiny web-app for administering and scoring the computer adaptive Philadelphia Naming Test.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rbcavanaugh.github.io/acom/}
  \item \url{https://github.com/rbcavanaugh/acom}
  \item Report bugs at \url{https://github.com/rbcavanaugh/acom/issues}
}

}
\author{
\strong{Maintainer}: Robert Cavanaugh \email{rob.cavanaugh@pitt.edu}

Authors:
\itemize{
  \item William D. Hula
}

}
\keyword{internal}
