% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_acom.R
\name{acom_tab_div}
\alias{acom_tab_div}
\title{Shows the user interface for the practice slides.}
\usage{
acom_tab_div(values)
}
\arguments{
\item{values}{values}
}
\description{
Shows the user interface for the practice slides.
}
