% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_acom_files.R
\name{combine_acom_files}
\alias{combine_acom_files}
\title{Combine ACOM files}
\usage{
combine_acom_files(folder_path, summary = FALSE)
}
\arguments{
\item{folder_path}{a path to a folder containing only unmodified acom output files}

\item{summary}{whether to return a summary of the final scores for each file. defaults to FALSE}
}
\value{
a .csv file of all data or summary data
}
\description{
This function returns either all responses combined or a summary table of final ability and sem scores
given a folder with only unmodified output of files generated from the ACOM 4.0.0 application.
}
