% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{downloadReportServer}
\alias{downloadReportServer}
\title{Download report module}
\usage{
downloadReportServer(id, values, name, notes)
}
\arguments{
\item{id}{}

\item{values}{}

\item{name}{}

\item{notes}{}
}
\description{
Download report module
}
