% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_results.R
\name{downloadResultsServer}
\alias{downloadResultsServer}
\title{Download results module}
\usage{
downloadResultsServer(id, values, in_progress)
}
\arguments{
\item{id}{}

\item{values}{}

\item{in_progress}{}
}
\description{
Download results module
}
