% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caption.R
\name{get_caption}
\alias{get_caption}
\title{The caption for the plot.}
\usage{
get_caption(values)
}
\arguments{
\item{values}{all values}
}
\value{
a html tag with the caption in italics
}
\description{
It changes based on whether its a repeat administration or not.
}
