% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_download.R
\name{get_data_for_download}
\alias{get_data_for_download}
\title{Compiles data for downloading}
\usage{
get_data_for_download(values, in_progress)
}
\arguments{
\item{values}{all the reactive values in the session}

\item{in_progress}{whether or not the test was ended early.}
}
\value{
a dataframe of responses for downloading
}
\description{
Compiles data for downloading
}
