% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_numbers.R
\name{get_final_numbers}
\alias{get_final_numbers}
\title{obtains final numbers from the testing data.}
\usage{
get_final_numbers(out, previous, num_previous)
}
\arguments{
\item{out}{list of the final numbers from the get_irt function}

\item{previous}{previous items}

\item{num_previous}{how many previous tests...0 or 1}
}
\value{
a small dataframe of final number estimates
}
\description{
obtains final numbers from the testing data.
}
