% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_item.R
\name{get_first_item}
\alias{get_first_item}
\title{get first item}
\usage{
get_first_item(all_items, previous, exclude_previous = F)
}
\arguments{
\item{all_items}{out}

\item{previous}{prev}

\item{exclude_previous}{num prev}
}
\description{
get first item
}
