% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_warning.R
\name{get_item_warning}
\alias{get_item_warning}
\title{Item Warning
returns a warning if not enough items are administered}
\usage{
get_item_warning(values)
}
\value{
a warning div
}
\description{
Item Warning
returns a warning if not enough items are administered
}
