% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_summary.R
\name{get_text_summary}
\alias{get_text_summary}
\title{Text summary for the top of the results page above the plot}
\usage{
get_text_summary(v)
}
\arguments{
\item{v}{all data}
}
\description{
Text summary for the top of the results page above the plot
}
