% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_time.R
\name{get_user_time}
\alias{get_user_time}
\title{Get user time
returns the users time so that the app can correctly
timestamp files}
\usage{
get_user_time()
}
\value{
the users datetime
}
\description{
Get user time
returns the users time so that the app can correctly
timestamp files
}
