% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_irt_function.R
\name{irt_function}
\alias{irt_function}
\title{irt magic function}
\usage{
irt_function(
  all_items,
  IRT = T,
  exclude_previous = F,
  previous,
  test = NA,
  exclude_eskimo = T
)
}
\arguments{
\item{all_items}{everything}

\item{IRT}{irt yes no}

\item{exclude_previous}{like name}

\item{previous}{prev if}

\item{test}{test}
}
\description{
irt magic function
}
