% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_practice.R
\name{practice_tab_div}
\alias{practice_tab_div}
\title{Shows the user interface for the practice slides.}
\usage{
practice_tab_div(values)
}
\arguments{
\item{values}{values}
}
\description{
Shows the user interface for the practice slides.
}
