% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_calculations.R
\name{response_to_numeric}
\alias{response_to_numeric}
\title{response_to_numeric}
\usage{
response_to_numeric(select, clarify, merge)
}
\arguments{
\item{select}{selected response}

\item{clarify}{ask for clarification for doesn't apply}
}
\value{
single row dataframe with response, response_num, response_merge, and clarify
}
\description{
response_to_numeric
}
