% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_results_upload.R
\name{results_tab_div2}
\alias{results_tab_div2}
\title{Shows the user interface for the resutls page when using the rescore function}
\usage{
results_tab_div2()
}
\description{
This is a different page for reasons I can't recall right now...
}
