% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_uploaded_data.R
\name{score_uploaded_data}
\alias{score_uploaded_data}
\title{Scores data that has been uploaded.}
\usage{
score_uploaded_data(values)
}
\value{
a list containing final IRT estimates and a text summary via get_text_summary()
}
\description{
Scores data that has been uploaded.
}
