% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_slides.R
\name{slides_tab_div}
\alias{slides_tab_div}
\title{User interface for the assessment tab}
\usage{
slides_tab_div(values)
}
\arguments{
\item{values}{values}

\item{progbar}{progress bar (not used)}
}
\description{
Shows PNT images during testing (not practice). Also
shows the 1 and 2 in the top right of the screen.
}
