% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_functions.R
\name{uploadData}
\alias{uploadData}
\title{Upload previous data}
\usage{
uploadData(file_input, rescore = F)
}
\arguments{
\item{file_input}{}
}
\value{
a dataframe of previous scores, or an empty df if not validated.
}
\description{
Upload previous data
}
