% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_download.R
\name{get_data_for_download}
\alias{get_data_for_download}
\title{gets data for download.}
\usage{
get_data_for_download(dat)
}
\arguments{
\item{dat}{}
}
\value{
a dataframe of responses
}
\description{
gets data for download.
}
