% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_numbers.R
\name{get_final_numbers}
\alias{get_final_numbers}
\title{final numbers}
\usage{
get_final_numbers(out, previous, num_previous)
}
\arguments{
\item{out}{out}

\item{previous}{prev}

\item{num_previous}{num prev}
}
\description{
final numbers
}
