% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_item.R
\name{get_first_item}
\alias{get_first_item}
\title{obtains the first item for the test}
\usage{
get_first_item(all_items, previous, exclude_previous = F)
}
\arguments{
\item{all_items}{at this stage, a dataframe of all possible items}

\item{previous}{are there previous items to exclude}

\item{exclude_previous}{whether or not to exclude those previous items}
}
\value{
a slide number for the first item
}
\description{
because the tests are different and have different first items,
}
