% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_data_long.R
\name{get_results_data_long}
\alias{get_results_data_long}
\title{Title Get Results data long}
\usage{
get_results_data_long(values)
}
\arguments{
\item{values}{list of stuff from server}
}
\description{
Title Get Results data long
}
