% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_summary.R
\name{get_text_summary}
\alias{get_text_summary}
\title{get text summary}
\usage{
get_text_summary(ability, sem, last_ability, last_sem, num_previous)
}
\arguments{
\item{ability}{ability}

\item{last_ability}{last abil}

\item{num_previous}{num prev}

\item{acc}{acc}

\item{ci_95}{ci95}

\item{last_ci_95}{last ci95}

\item{first_ability}{last first}

\item{first_ci_95}{first ci}
}
\description{
get text summary
}
