% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_column.R
\name{pull_column}
\alias{pull_column}
\title{Title Pull Column}
\usage{
pull_column(dat, pull)
}
\arguments{
\item{dat}{dataframe in}

\item{pull}{column to select}
}
\value{

}
\description{
Title Pull Column
}
