% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_slides.R
\name{slides_tab_div}
\alias{slides_tab_div}
\title{slides tab}
\usage{
slides_tab_div(values, progbar)
}
\arguments{
\item{values}{values}

\item{progbar}{progress bar (not used)}
}
\description{
slides tab
}
